uboot add de, bmp, auduo, pmu229

+++ sprite/firmware/imgdecode.h
@@ -68,7 +68,7 @@ extern   long long 		Img_GetItemSize	(HIMAGE hImage, HIMAGEITEM hItem);
 
 extern   uint 			Img_GetItemStart(HIMAGE hImage, HIMAGEITEM hItem);
 
-extern   int 			Img_ReadItem	(HIMAGE hImage, HIMAGEITEM hItem, void *buffer, long long buffer_size);
+extern   uint 			Img_ReadItem	(HIMAGE hImage, HIMAGEITEM hItem, void *buffer, uint buffer_size);
 
 extern   int 			Img_CloseItem	(HIMAGE hImage, HIMAGEITEM hItem);
